requirements.txt
amqp==2.4.2
billiard==3.6.0.0
boto3==1.9.121
botocore==1.12.121
celery==4.3.0
coverage==4.5.3
Django==2.1.7
django-money==0.14.4
django-storages==1.7.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
docutils==0.14
jmespath==0.9.4
kombu==4.5.0
Pillow==5.4.1
psycopg2==2.7.7
py-moneyed==0.8.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
vine==1.3.0
